Updated screens implementation to use class inheritance
